pyproject.toml
[project]
name = "ashkit"
version = "0.1.0"
description = "Lightweight personal AI assistant with Feishu, Memory, MCP and Skills"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "SQLAlchemy>=2.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
    "tenacity>=8.0",
    "python-dotenv>=1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[project.optional-dependencies]
all = [
    "openai>=1.0",
    "anthropic>=0.18",
]
feishu = [
    "lark>=1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]
